# Payments/openapi-spec.json
{"openapi":"3.0.0","info":{"title":"Payments API","version":"1.0.0","description":"Swagger docs for Payments microservice"},"servers":[{"url":"http://localhost:4003"}],"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}},"security":[{"bearerAuth":[]}],"paths":{"/api/payments/place-order":{"post":{"summary":"Place a new order transaction","description":"Creates a new transaction for an event and returns a transaction_id.","tags":["Payments"],"requestBody":{"required":true,"content":{"application/json":{"example":{"eventId":101}}}},"responses":{"201":{"description":"Transaction created successfully","content":{"application/json":{"example":{"success":true,"message":"Transaction created successfully","transaction_id":"550e8400-e29b-41d4-a716-446655440000","event_details":{"id":101,"name":"Sample Event"}}}}},"400":{"description":"Missing eventId","content":{"application/json":{"example":{"success":false,"message":"Event ID is required to place an order."}}}},"404":{"description":"Event not found","content":{"application/json":{"example":{"success":false,"message":"Event not found."}}}},"500":{"description":"Transaction creation failed","content":{"application/json":{"example":{"success":false,"message":"Transaction creation failed"}}}}}}},"/api/payments/create-order":{"post":{"summary":"Create an order in Razorpay","description":"Creates an order in Razorpay using transaction details.","tags":["Payments"],"requestBody":{"required":true,"content":{"application/json":{"example":{"transaction_id":"550e8400-e29b-41d4-a716-446655440000","amount":500,"currency":"INR","organizationId":"org_001","eventId":101}}}},"responses":{"200":{"description":"Order created successfully","content":{"application/json":{"example":{"success":true,"order_id":"order_JDfklw93u2","transaction_id":"550e8400-e29b-41d4-a716-446655440000","amount":500,"currency":"INR","notes":{"organizationId":"org_001","eventId":101}}}}},"400":{"description":"Missing parameters","content":{"application/json":{"example":{"success":false,"message":"Missing required parameters (transaction_id, amount, currency, organizationId, evenId)"}}}},"500":{"description":"Order creation failed","content":{"application/json":{"example":{"success":false,"message":"Order creation failed"}}}}}}},"/api/payments/generate-signature":{"post":{"summary":"Generate payment signature","description":"Generates a signature to verify payment authenticity.","tags":["Payments"],"requestBody":{"required":true,"content":{"application/json":{"example":{"razorpay_order_id":"order_JDfklw93u2","razorpay_payment_id":"pay_29QQoUBi66xm2f","payment_status":"captured"}}}},"responses":{"200":{"description":"Signature generated successfully","content":{"application/json":{"example":{"success":true,"signature":"5d5c5f93c0f5f91f..."}}}},"400":{"description":"Missing parameters","content":{"application/json":{"example":{"success":false,"message":"Missing payment verification parameters"}}}},"500":{"description":"Signature generation error","content":{"application/json":{"example":{"success":false,"message":"Signature generation error"}}}}}}},"/api/payments/order-status/{order_id}":{"post":{"summary":"Get order status","description":"Fetches the current payment/order status from Razorpay.","tags":["Payments"],"parameters":[{"in":"path","name":"order_id","required":true,"schema":{"type":"string"},"description":"Razorpay order ID"}],"responses":{"200":{"description":"Order status fetched successfully","content":{"application/json":{"example":{"success":true,"order_id":"order_JDfklw93u2","payment_status":"paid","fetched_at":"04-10-2025 14:35:22"}}}},"400":{"description":"Missing order ID","content":{"application/json":{"example":{"success":false,"message":"Order ID is required"}}}},"500":{"description":"Failed to fetch order status","content":{"application/json":{"example":{"success":false,"message":"Failed to fetch order status"}}}}}}}},"tags":[]}